+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [0.1.5] - 2025.03.07
 
+### Added
+- Checksum verification for BMS data response
+
 ### Changed
 - Updated minor version of `dbus-fast` library
 - Code formatting